requirements.txt
numpy
pandas
librosa
soundfile
pytorch
tensorflow